requirements/local.txt
-r base.txt
django-extensions>=3.2.3
pre-commit==3.6.2
pytest-django>=4.7.0

